+++ Changes.md
@@ -11,6 +11,7 @@ API
 ---
 
 - MenuButton : Added `set/getImmediate()` methods. These allow the button to execute a menu item immediately instead of showing the menu, if the menu has only a single item.
+- Signals : Added CatchingCombiner Python class, equivalent to the C++ template class.
 
 1.6.6.1 (relative to 1.6.6.0)
 =======